fixes loading bug
Now the cards are 20 per page as expected.
